pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctyun-cli"
version = "1.3.7"
description = "天翼云CLI工具 - 基于终端的云资源管理平台"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Y.FENG", email = "popfrog@gmail.com"}
]
maintainers = [
    {name = "Y.FENG", email = "popfrog@gmail.com"}
]
keywords = [
    "ctyun", "cloud", "cli", "management", "monitoring", "ecs",
    "redis", "distributed-cache", "query", "snapshot", "keypair",
    "volume", "backup", "affinity-group", "flavor", "resize",
    "vnc", "statistics", "api", "devops"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Topic :: Internet :: WWW/HTTP",
    "Environment :: Console",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "build>=0.10.0",
    "twine>=4.0.0"
]
test = ["pytest>=7.4.0", "pytest-cov>=4.1.0"]
lint = ["black>=23.0.0", "flake8>=6.0.0"]
build = ["build>=0.10.0", "twine>=4.0.0"]

[project.urls]
Homepage = "https://pypi.org/project/ctyun-cli/"
Documentation = "https://github.com/fengyucn/ctyun-cli"
Repository = "https://github.com/fengyucn/ctyun-cli"
"Bug Tracker" = "https://github.com/fengyucn/ctyun-cli/issues"
Changelog = "https://github.com/fengyucn/ctyun-cli/commits/master"

[project.scripts]
ctyun-cli = "cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"